+++ CHANGELOG.md
@@ -4,8 +4,16 @@ All notable changes to the "vscode-splunk-search-linter" extension will be docum
 
 ## [Unreleased]
 
-## [1.0.0] - 2020-01-09
+## [0.3.0] - 2020-01-22
+
+### Changed
+
+- Splunk Language ID to `splunk_search` for compatibility with Splunk search language to fix issue #2
+
+## [0.1.0] - 2020-01-09
+
 ### Added
+
 - Autocompleting commands
 - Syntax for main, eval, and stats commands
 - Description for commands